+++ Changelog.md
@@ -225,6 +225,7 @@ Formio.createForm(document.getElementById('formio'), 'https://examples.form.io/e
  - FIO-7334: Fixes an issue where Radio values do not appear for Action Conditions settings
  - FIO-8009: fixed display of the required asterisk
  - FIO-8111: fixed saveDraft Trigger for nested forms
+ - FIO-8109: fixed save draft triggering for nested components
 
 ## 5.0.0-rc.37
 ### Fixed